Move KB into subdirectory.
