+++ spacy/kb/__init__.py
@@ -1,0 +1,2 @@
+from kb import KnowledgeBase
+from kb_in_memory_lookup import InMemoryLookupKB
