+++ spacy/kb/kb_in_memory_lookup.pyx
