+++ spacy/tests/serialize/test_resource_warning.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 import pytest
 import srsly
 from numpy import zeros
-from spacy.kb_inmemorylookup import InMemoryLookupKB, Writer
+from spacy.kb_in_memory_lookup import InMemoryLookupKB, Writer
 from spacy.vectors import Vectors
 from spacy.language import Language
 from spacy.pipeline import TrainablePipe
